API Access Token

+++ src/pages/api/[[...user]].ts
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/services";
 import type { NextApiRequest, NextApiResponse } from "next";
+import jwt from "jsonwebtoken";
 
 export default async function handler(
   req: NextApiRequest,
@@ -15,35 +16,57 @@ export default async function handler(
       .status(200)
       .json({ status: true, statusCode: 200, message: "success", data });
   } else if (req.method === "PUT") {
-    const { id, data } = req.body;
-    await updateData(
-      "users",
-      "TCsvDHq8AoYfbyr00Vn5",
-      data,
-      (result: boolean) => {
-        if (result) {
-          res
-            .status(200)
-            .json({ status: true, statusCode: 200, message: "success" });
+    const { data } = req.body;
+    const { user }: any = req.query;
+    const accessToken = req.headers.authorization?.split(" ")[1] || "";
+    jwt.verify(
+      accessToken,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        if (decoded && decoded.role === "admin") {
+          await updateData("users", user[1], data, (result: boolean) => {
+            if (result) {
+              res
+                .status(200)
+                .json({ status: true, statusCode: 200, message: "success" });
+            } else {
+              res
+                .status(400)
+                .json({ status: false, statusCode: 400, message: "failed" });
+            }
+          });
         } else {
           res
-            .status(400)
-            .json({ status: false, statusCode: 400, message: "failed" });
+            .status(403)
+            .json({ status: true, statusCode: 403, message: "access denied" });
         }
       }
     );
   } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    await deleteData("users", user[1], (result: boolean) => {
-      if (result) {
-        res
-          .status(200)
-          .json({ status: true, statusCode: 200, message: "success" });
-      } else {
-        res
-          .status(400)
-          .json({ status: false, statusCode: 400, message: "failed" });
+    const accessToken = req.headers.authorization?.split(" ")[1] || "";
+    jwt.verify(
+      accessToken,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        if (decoded && decoded.role === "admin") {
+          await deleteData("users", user[1], (result: boolean) => {
+            if (result) {
+              res
+                .status(200)
+                .json({ status: true, statusCode: 200, message: "success" });
+            } else {
+              res
+                .status(400)
+                .json({ status: false, statusCode: 400, message: "failed" });
+            }
+          });
+        } else {
+          res
+            .status(403)
+            .json({ status: true, statusCode: 403, message: "access denied" });
+        }
       }
-    })
+    );
   }
 }
